pyproject.toml
[tool.poetry]
name = "apptrix_test"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.3"
psycopg2-binary = "^2.9.3"
djangorestframework = "^3.13.1"
django-filter = "^21.1"
Pillow = "^9.0.1"
loguru = "^0.6.0"
django-extensions = "^3.1.5"
djangorestframework-gis = "^0.18"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
